requirements.txt
Django==4.2
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9  # If using PostgreSQL
whitenoise==6.6.0  # For serving static files in production
ffmpeg-python==0.2.0
python-dotenv==1.0.1  # For managing environment variables
requests==2.31.0
daphne==4.0.0  # ASGI server for Django Channels
python-crontab==3.2.0  # Required for scheduled tasks
